backend/requirements.txt
fastapi[standard]==0.115.5
python-dotenv==1.0.1
google-generativeai==0.8.3
Pillow==11.0.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.9
httpx==0.27.2

